Added udp extension files, enabled csensors and other files as well

+++ bsp/boards/nrf52840/adc_sensor.h
@@ -20,8 +20,8 @@ bool adc_sens_init(void);
 uint16_t adc_sens_read_battery(void);
 float adc_sens_convert_battery(uint16_t raw);
 
-uint16_t adc_sens_read_temperature(void);
-float adc_sens_convert_temperature(uint16_t cpu_temp_raw);
+int32_t adc_sens_read_temperature(void);
+float adc_sens_convert_temperature(int32_t cpu_temp_raw);
 
 
 #endif // __ADC_SENSOR_H__
